feat(DTFS2-7578): move services into clone folder
