+++ portal-api/src/v1/gef/controllers/clone-gef-deal.controller.js
@@ -1,8 +1,8 @@
 const { ApiError } = require('@ukef/dtfs2-common');
 
 const { generatePortalAuditDetails } = require('@ukef/dtfs2-common/change-stream');
-const { cloneDealToLatestVersion } = require('../services/clone-deal.service');
-const { cloneFacilities } = require('../services/clone-facilities.service');
+const { cloneDealToLatestVersion } = require('../services/clone/clone-deal.service');
+const { cloneFacilities } = require('../services/clone/clone-facilities.service');
 const { validateApplicationReferences } = require('./validation/application');
 
 exports.clone = async (req, res) => {
